return (float)

+++ main.c
@@ -1,18 +1,18 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-float tinhNuaChuViTamGiac(float a, float  b, float c){
-    return (a + b + c) / 2;
+float tinhNuaChuViTamGiac(int a, int b, int c){
+    return (float) (a + b + c) / 2;
 }
 int main() {
 
-    float a, b, c;
+    int a, b, c;
     printf("Nhap so do canh thu nhat: ");
-    scanf("%f", &a);
+    scanf("%d", &a);
     printf("Nhap so do canh thu hai: ");
-    scanf("%f", &b);
+    scanf("%d", &b);
     printf("Nhap so do canh thu ba: ");
-    scanf("%f", &c);
+    scanf("%d", &c);
     if (a > 0 && b > 0 && c > 0){
         if ((a+b>c) && (a+c>b) && (b+c>a)) {
             printf("Nua chu vi cua tam giac la %.2f", tinhNuaChuViTamGiac(a, b, c));
